:sparkles: move generate3dShadow to utils file
