+++ src/components/hero.js
@@ -14,14 +14,6 @@ const heroStyle = css`
     height: 65rem;
 `
 
-export const generate3dShadow = (size, color) => {
-  let result = []
-  for (let i = size; i > 0; i -= 1) {
-    result = [...result, `${i}px ${i}px ${color}`]
-  }
-  return result.join()
-}
-
 const Hero = () => (
   <Box css={heroStyle}>
     <Box
