+++ src/utils.js
@@ -1,0 +1,8 @@
+/* eslint-disable import/prefer-default-export */
+export const generate3dShadow = (size, color) => {
+  let result = []
+  for (let i = size; i > 0; i -= 1) {
+    result = [...result, `${i}px ${i}px ${color}`]
+  }
+  return result.join()
+}
